refactor(components): refactor Dialog and update imports and exports
fix #307
