+++ packages/components/src/base/Backdrop/Backdrop.tsx
@@ -1,0 +1,5 @@
+import { Backdrop as MuiBackdrop, type BackdropProps as MuiBackdropProps } from '@mui/material';
+
+export function Backdrop(props: MuiBackdropProps): JSX.Element {
+  return <MuiBackdrop {...props} />;
+}
