+++ packages/components/src/base/Backdrop/index.tsx
@@ -1,0 +1,1 @@
+export { Backdrop } from './Backdrop';
