+++ packages/components/src/base/Dialog/Dialog.tsx
@@ -1,0 +1,5 @@
+import { Dialog as MuiDialog, type DialogProps as MuiDialogProps } from '@mui/material';
+
+export function Dialog(props: MuiDialogProps): JSX.Element {
+  return <MuiDialog {...props} />;
+}
