--- packages/components/src/base/Dialog/backdrop.tsx
@@ -1,5 +1,0 @@
-import { Backdrop as MuiBackdrop, type BackdropProps } from '@mui/material';
-
-export function Backdrop(props: BackdropProps): JSX.Element {
-  return <MuiBackdrop {...props} />;
-}
