--- packages/components/src/base/Dialog/dialog.tsx
@@ -1,5 +1,0 @@
-import { Dialog as MuiDialog, type DialogProps } from '@mui/material';
-
-export function Dialog(props: DialogProps): JSX.Element {
-  return <MuiDialog {...props} />;
-}
