--- packages/components/src/base/Dialog/dialogactions.tsx
@@ -1,5 +1,0 @@
-import { DialogActions as MuiDialogActions, type DialogActionsProps } from '@mui/material';
-
-export function DialogActions(props: DialogActionsProps): JSX.Element {
-  return <MuiDialogActions {...props} />;
-}
