--- packages/components/src/base/Dialog/dialogcontent.tsx
@@ -1,5 +1,0 @@
-import { DialogContent as MuiDialogContent, type DialogContentProps } from '@mui/material';
-
-export function DialogContent(props: DialogContentProps): JSX.Element {
-  return <MuiDialogContent {...props} />;
-}
