--- packages/components/src/base/Dialog/dialogcontenttext.tsx
@@ -1,8 +1,0 @@
-import {
-  DialogContentText as MuiDialogContentText,
-  type DialogContentTextProps
-} from '@mui/material';
-
-export function DialogContentText(props: DialogContentTextProps): JSX.Element {
-  return <MuiDialogContentText {...props} />;
-}
