--- packages/components/src/base/Dialog/dialogtitle.tsx
@@ -1,5 +1,0 @@
-import { DialogTitle as MuiDialogTitle, type DialogTitleProps } from '@mui/material';
-
-export function DialogTitle(props: DialogTitleProps): JSX.Element {
-  return <MuiDialogTitle {...props} />;
-}
