+++ packages/components/src/base/Dialog/index.tsx
@@ -1,6 +1,1 @@
-export { Backdrop } from './backdrop';
-export { Dialog } from './dialog';
-export { DialogActions } from './dialogactions';
-export { DialogContent } from './dialogcontent';
-export { DialogContentText } from './dialogcontenttext';
-export { DialogTitle } from './dialogtitle';
+export { Dialog } from './Dialog';
