+++ packages/components/src/base/DialogActions/DialogActions.tsx
@@ -1,0 +1,8 @@
+import {
+  DialogActions as MuiDialogActions,
+  type DialogActionsProps as MuiDialogActionsProps
+} from '@mui/material';
+
+export function DialogActions(props: MuiDialogActionsProps): JSX.Element {
+  return <MuiDialogActions {...props} />;
+}
