+++ packages/components/src/base/DialogActions/index.tsx
@@ -1,0 +1,1 @@
+export { DialogActions } from './DialogActions';
