+++ packages/components/src/base/DialogContent/DialogContent.tsx
@@ -1,0 +1,8 @@
+import {
+  DialogContent as MuiDialogContent,
+  type DialogContentProps as MuiDialogContentProps
+} from '@mui/material';
+
+export function DialogContent(props: MuiDialogContentProps): JSX.Element {
+  return <MuiDialogContent {...props} />;
+}
