+++ packages/components/src/base/DialogContent/index.tsx
@@ -1,0 +1,1 @@
+export { DialogContent } from './DialogContent';
