+++ packages/components/src/base/DialogContentText/DialogContentText.tsx
@@ -1,0 +1,8 @@
+import {
+  DialogContentText as MuiDialogContentText,
+  type DialogContentTextProps as MuiDialogContentTextProps
+} from '@mui/material';
+
+export function DialogContentText(props: MuiDialogContentTextProps): JSX.Element {
+  return <MuiDialogContentText {...props} />;
+}
