+++ packages/components/src/base/DialogContentText/index.tsx
@@ -1,0 +1,1 @@
+export { DialogContentText } from './DialogContentText';
