+++ packages/components/src/base/DialogTitle/DialogTitle.tsx
@@ -1,0 +1,8 @@
+import {
+  DialogTitle as MuiDialogTitle,
+  type DialogTitleProps as MuiDialogTitleProps
+} from '@mui/material';
+
+export function DialogTitle(props: MuiDialogTitleProps): JSX.Element {
+  return <MuiDialogTitle {...props} />;
+}
