+++ packages/components/src/base/DialogTitle/index.tsx
@@ -1,0 +1,1 @@
+export { DialogTitle } from './DialogTitle';
