+++ packages/components/src/custom/Dialog/index.tsx
@@ -1,6 +1,4 @@
-import StyledDialog from './StyledDialog';
-import StyledDialogActions from './StyledDialogActions';
-import StyledDialogContent from './StyledDialogContent';
-import StyledDialogTitle from './StyledDialogTitle';
-
-export { StyledDialog, StyledDialogActions, StyledDialogContent, StyledDialogTitle };
+export { default as StyledDialog } from './StyledDialog';
+export { default as StyledDialogActions } from './StyledDialogActions';
+export { default as StyledDialogContent } from './StyledDialogContent';
+export { default as StyledDialogTitle } from './StyledDialogTitle';
