+++ packages/components/src/custom/index.tsx
@@ -1,16 +1,8 @@
 // import { ErrorBoundary, withErrorBoundary, withSuppressedErrorBoundary } from './ErrorBoundary';
-import { StyledDialog, StyledDialogActions, StyledDialogContent } from './Dialog';
 import { useWindowDimensions } from './Helpers/Dimension';
 import { useNotificationHandler } from './Helpers/Notification';
 import { StyledTooltip } from './Tooltip';
 
 export { StyledChartDialog } from './ChartDialog';
 export { StyledSearchBar } from './StyledSearchBar';
-export {
-  StyledDialog,
-  StyledDialogActions,
-  StyledDialogContent,
-  StyledTooltip,
-  useNotificationHandler,
-  useWindowDimensions
-};
+export { StyledTooltip, useNotificationHandler, useWindowDimensions };
